pyproject.toml
[project]
name = "metcs-633-bu-agent"
version = "0.1.0"
description = "An agent designed for BU MET CS633 (Fall 2025). Its intent is to provide students with career advice and course recommendations based on the tent."
authors = [
    {name = "Sean Silver", email = "smsilver@bu.edu"},
    {name = "Brendan Collins", email = "bcollins@bu.edu"},
    {name = "Gagan Veginati", email = "gaganvbu@bu.edu"},
    {name = "Jonthan Rothendler", email = "jroth@bu.edu"},
    {name = "Zhara Omirzak", email = "zharas01@bu.edu"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"

dependencies = [
    "black>=25.9.0",
    "docker>=7.1.0",
    "firestore>=0.0.8",
    "flask>=3.1.2",
    "google-adk>=1.15.1",
    "kubernetes>=34.1.0",
    "litellm>=1.77.7",
    "logging>=0.4.9.6",
    "networkx>=3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "asyncio>=4.0.0",
    # front-end packages
    "react>=4.3.0",
    "spacy>=3.8.7",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    # pytest packages
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-trio>=0.8.0",
    "pytest-twisted>=1.14.3",
    "pytest-tornasync>=0.6.0.post2",
    "twisted>=25.5.0",
    "anyio>=4.11.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [
  "./src/"
]
